requirements.txt.txt
streamlit
mysql-connector-python
pandas
pymupdf